% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{all_check_groups}
\alias{all_check_groups}
\title{List available check group names}
\usage{
all_check_groups()
}
\value{
A character vector of check group names.
}
\description{
Returns the names of all registered check groups.
Use these names with \code{\link[=checks_by_group]{checks_by_group()}} to select checks by group,
or with \code{options(goodpractice.exclude_check_groups = ...)} to skip
groups.
Full descriptions of each check group are return by
\code{\link[=describe_check_groups]{describe_check_groups()}}.
}
\examples{
# Names of all check groups:
all_check_groups()

# List individual checks by group:
chks <- lapply(all_check_groups(), checks_by_group)
names(chks) <- all_check_groups()
chks
}
\seealso{
Other check_groups:
\code{\link[=all_checks]{all_checks()}},
\code{\link[=checks_by_group]{checks_by_group()}},
\code{\link[=default_checks]{default_checks()}},
\code{\link[=describe_check]{describe_check()}},
\code{\link[=describe_check_groups]{describe_check_groups()}},
\code{\link[=tidyverse_checks]{tidyverse_checks()}}
}
\concept{check_groups}
