% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{default_checks}
\alias{default_checks}
\title{List the names of default checks (excludes optional check sets)}
\usage{
default_checks()
}
\value{
Character vector of default check names
}
\description{
List the names of default checks (excludes optional check sets)
}
\examples{
default_checks()
}
\seealso{
Other check_groups:
\code{\link[=all_check_groups]{all_check_groups()}},
\code{\link[=all_checks]{all_checks()}},
\code{\link[=checks_by_group]{checks_by_group()}},
\code{\link[=describe_check]{describe_check()}},
\code{\link[=describe_check_groups]{describe_check_groups()}},
\code{\link[=tidyverse_checks]{tidyverse_checks()}}
}
\concept{check_groups}
