% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_r_file_pattern}
\alias{default_r_file_pattern}
\title{Default pattern for R files}
\usage{
default_r_file_pattern()
}
\value{
Regular expression.
}
\description{
Default pattern for R files
}
\keyword{internal}
