% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{describe_check}
\alias{describe_check}
\title{Describe one or more checks}
\usage{
describe_check(check_name = NULL)
}
\arguments{
\item{check_name}{Names of checks to be described.}
}
\value{
List of character descriptions for each \code{check_name}
}
\description{
Describe one or more checks
}
\examples{
describe_check("rcmdcheck_non_portable_makevars")
check_name <- c("no_description_depends",
                "lintr_assignment_linter",
                "no_import_package_as_a_whole",
                "rcmdcheck_missing_docs")
describe_check(check_name)
# Or to see all checks:
\dontrun{
  describe_check(all_checks())
}
}
\seealso{
Other check_groups:
\code{\link[=all_check_groups]{all_check_groups()}},
\code{\link[=all_checks]{all_checks()}},
\code{\link[=checks_by_group]{checks_by_group()}},
\code{\link[=default_checks]{default_checks()}},
\code{\link[=describe_check_groups]{describe_check_groups()}},
\code{\link[=tidyverse_checks]{tidyverse_checks()}}
}
\concept{check_groups}
