% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{describe_check_groups}
\alias{describe_check_groups}
\title{Describe available check groups}
\usage{
describe_check_groups()
}
\value{
A named list of each check group defined in \code{\link[=all_check_groups]{all_check_groups()}},
with text descriptions of each group.
}
\description{
Returns full descriptions of all registered check groups.
}
\examples{
# Names of all check groups:
all_check_groups()

# And corresponding descriptions:
describe_check_groups()
}
\seealso{
Other check_groups:
\code{\link[=all_check_groups]{all_check_groups()}},
\code{\link[=all_checks]{all_checks()}},
\code{\link[=checks_by_group]{checks_by_group()}},
\code{\link[=default_checks]{default_checks()}},
\code{\link[=describe_check]{describe_check()}},
\code{\link[=tidyverse_checks]{tidyverse_checks()}}
}
\concept{check_groups}
