% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_json}
\alias{export_json}
\title{Export failed checks to JSON}
\usage{
export_json(gp, file, pretty = FALSE)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}

\item{file}{Output connection or file.}

\item{pretty}{Whether to pretty-print the JSON.}
}
\value{
Invisibly returns the path to the output file.
}
\description{
Export failed checks to JSON
}
\examples{
path <- system.file("bad1", package = "goodpractice")
g <- gp(path, checks = "description_url")
tmp <- tempfile(fileext = ".json")
export_json(g, tmp)
unlink(tmp)
}
\seealso{
Other output:
\code{\link[=print.goodPractice]{print.goodPractice()}},
\code{\link[=results]{results()}}
}
\concept{output}
