% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{failed_checks}
\alias{failed_checks}
\title{Names of the failed checks}
\usage{
failed_checks(gp)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}
}
\value{
Names of the failed checks.
}
\description{
Names of the failed checks
}
\examples{
path <- system.file("bad1", package = "goodpractice")
# run a subset of all checks available
g <- gp(path, checks = all_checks()[9:16])
failed_checks(g)
# Or run with named check groups
g <- gp(path, checks = checks_by_group("description", "namespace"))
failed_checks(g)
}
\seealso{
Other API:
\code{\link[=checks]{checks()}},
\code{\link{customization}},
\code{\link[=failed_positions]{failed_positions()}}
}
\concept{API}
