% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{failed_positions}
\alias{failed_positions}
\title{Positions of check failures in the source code}
\usage{
failed_positions(gp)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}
}
\value{
A list of lists of positions. See details below.
}
\description{
Note that not all checks refer to the source code.
For these the result will be \code{NULL}.
}
\details{
For the ones that do, the results is a list, one for each failure.
Since the same check can fail multiple times. A single failure
is a list with entries: \code{filename}, \code{line_number},
\code{column_number}, \code{ranges}. \code{ranges} is a list of
pairs of start and end positions for each line involved in the
check.
}
\examples{
path <- system.file("bad1", package = "goodpractice")
g <- gp(path, checks = "description_url")
failed_positions(g)
}
\seealso{
Other API:
\code{\link[=checks]{checks()}},
\code{\link{customization}},
\code{\link[=failed_checks]{failed_checks()}}
}
\concept{API}
