% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_markers.R
\name{get_marker}
\alias{get_marker}
\title{Get a marker from the positions of a check}
\usage{
get_marker(gp, check)
}
\arguments{
\item{gp}{\code{gp()} output}

\item{check}{name of the check to extract}
}
\description{
Get a marker from the positions of a check
}
\keyword{internal}
