% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{learn_gp_skill}
\alias{learn_gp_skill}
\title{Function for AI agents to run and learn how to use the 'goodpractice'
package.}
\usage{
learn_gp_skill()
}
\value{
The content of the file \code{system.file("skills",
"goodpractice4agents.md", package = "goodpractice"))}
}
\description{
Function for AI agents to run and learn how to use the 'goodpractice'
package.
}
\examples{
\dontrun{
learn_gp_skill()
}
}
