% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customization.R
\name{make_check}
\alias{make_check}
\title{Create a check function}
\usage{
make_check(description, check, gp, ...)
}
\arguments{
\item{description}{A description of the check.}

\item{check}{A function that takes the \code{state} as an argument.}

\item{gp}{A short description of what is good practice.}

\item{...}{Further arguments.}
}
\description{
Create a check function
}
