% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customization.R
\name{make_prep}
\alias{make_prep}
\title{Create a preparation function}
\usage{
make_prep(name, func)
}
\arguments{
\item{name}{Name of the preparation function.}

\item{func}{A function that takes two arguments:
The \code{path} to the root directory of the package, and
a logical argument: \code{quiet}. If \code{quiet} is true,
the preparation function may print out diagnostic messages.}
}
\description{
Create a preparation function
}
