% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_rcmdcheck.R
\name{make_rcmd_check}
\alias{make_rcmd_check}
\title{Wrapper on make_check, specific to R CMD check}
\usage{
make_rcmd_check(
  description,
  pattern,
  gp = NULL,
  type = c("warnings", "notes", "errors"),
  tags = NULL,
  preps = NULL,
  ...
)
}
\arguments{
\item{description}{A description of the check.}

\item{pattern}{The text pattern identifying the check.}

\item{type}{Type of notification, one of "warnings", "notes" or "errors".}

\item{tags}{Tags to be passed on to make_check.}

\item{preps}{Preps to be passed on to make_check.}

\item{...}{Currently not supported.}
}
\description{
Wrapper on make_check, specific to R CMD check
}
\keyword{internal}
