% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{opt_in_checks}
\alias{opt_in_checks}
\title{List the names of opt-in checks}
\usage{
opt_in_checks()
}
\value{
Character vector of opt-in check names
}
\description{
These checks are included in \code{\link{all_checks}} but excluded from
\code{\link{default_checks}}. Include them individually by name or use
\code{checks = c(default_checks(), opt_in_checks())} to add all opt-in
checks to the default set.
}
