% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ goodpractice }
\alias{ goodpractice }
\alias{ goodpractice }
\title{ Good Practice Advice for Building R Packages }
\description{
Give advice about good practices when building R packages. Advice includes
  functions and syntax to avoid, package structure, code complexity, code
  formatting, etc.
}
