% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_expressions.R
\name{prep_expressions}
\alias{prep_expressions}
\title{Extract all closures from a package}
\usage{
prep_expressions(package, version = NULL)
}
\arguments{
\item{package}{Package name.}

\item{version}{Currently ignored.}
}
\value{
The closures in a named list.
}
\description{
The package must be extracted into the working directory, as usual.
}
\details{
We can use lintr to extract the functions, but need to use
our own code (based on similar code in functionMap) to
get the right collation order.
}
\keyword{internal}

