% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_expressions.R
\name{r_package_files}
\alias{r_package_files}
\title{Get all source files of a package, in the right order}
\usage{
r_package_files(path)
}
\arguments{
\item{path}{Path to the root of the R package.}
}
\value{
A character vector of (relative) file
  names in the corrent collation order.
}
\description{
It uses the `Collate` entry in the `DESCRIPTION` file,
if there is one. Otherwise the order is alphabetical.
}
\keyword{internal}
