% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{results}
\alias{results}
\title{Return all check results in a data frame}
\usage{
results(gp)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}
}
\value{
Data frame, with columns:
\item{check}{The name of the check.}
\item{passed}{Logical, whether the check passed.}
}
\description{
Return all check results in a data frame
}
\examples{
path <- system.file("bad1", package = "goodpractice")
# Run a subset of all checks available
g <- gp(path, checks = all_checks()[9:16])
results(g)
# Or run with named check groups
g <- gp(path, checks = checks_by_group("description", "namespace"))
results(g)
}
\seealso{
Other output:
\code{\link[=export_json]{export_json()}},
\code{\link[=print.goodPractice]{print.goodPractice()}}
}
\concept{output}
