% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{use_skill_gp}
\alias{use_skill_gp}
\title{Download a local copy of 'goodpractice4agents.md' to tell AI agents how to
use this package.}
\usage{
use_skill_gp(
  path = "goodpractice4agents.md",
  use_skills_subdir = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{path}{Local path where the file should be written. If \code{path} is
an existing directory, the file is written as \code{goodpractice4agents.md}
within it; otherwise \code{path} is treated as the full target file name.}

\item{use_skills_subdir}{If \code{TRUE}, the \code{path} will be edited to
write the file in to a \code{skills/} sub-directory.}

\item{overwrite}{If \code{FALSE} (default) and file specified by \code{path}
already exists, function will error and not overwrite the existing file.}
}
\value{
(Invisibly) the full path to the written file.
}
\description{
The 'goodpractice4agents.md' file contains sufficient instructions for most
AI agents to edit package code so that it passes most 'goodpractice'
checks. The file can be directly edited and tweaked for personal use cases.
An agent can be instructed to simply "run the file goodpractice4agents.md".
Alternatively, the file can be directly downloaded from GitHub at
\url{https://github.com/ropensci-review-tools/goodpractice/blob/main/inst/skills/goodpractice4agents.md}.
}
\details{
You can also instruct agents to call the companion function,
\link{learn_skill_gp}, directly in order to learn and use the skill itself.
}
\examples{
path <- file.path(tempdir(), "skill.md")
f <- use_skill_gp(path)
file.exists(f)
unlink(f)
}
