% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{all_check_groups}
\alias{all_check_groups}
\title{List available check group names}
\usage{
all_check_groups(describe = FALSE)
}
\arguments{
\item{describe}{If \code{TRUE}, print descriptions of all check groups to screen,
otherwise (default) just return the names of groups.}
}
\value{
For default 'describe = FALSE', a character vector of check group
names; for 'describe = TRUE', a named list with descriptions of all checks.
}
\description{
Returns the names, and optionally descriptions, of all registered check
groups.
Use these names with \code{\link[=checks_by_group]{checks_by_group()}} to select checks by group,
or with \code{options(goodpractice.exclude_check_groups = ...)} to skip
groups.
}
\examples{
# Default shows names of groups only:
all_check_groups()

# Show full group descriptions:
all_check_groups(describe = TRUE)

# List individual checks by group:
chks <- lapply(all_check_groups(), checks_by_group)
names(chks) <- all_check_groups()
chks
}
