% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{checks}
\alias{checks}
\title{List all checks performed}
\usage{
checks(gp)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}
}
\value{
Character vector of check names.
}
\description{
List all checks performed
}
\seealso{
Other API: \code{\link{failed_checks}},
  \code{\link{results}}
}
