% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{default_checks}
\alias{default_checks}
\title{List the names of default checks (excludes optional check sets)}
\usage{
default_checks()
}
\value{
Character vector of default check names
}
\description{
List the names of default checks (excludes optional check sets)
}
\examples{
default_checks()
}
