% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_json}
\alias{export_json}
\title{Export failed checks to JSON}
\usage{
export_json(gp, file, pretty = FALSE)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}

\item{file}{Output connection or file.}

\item{pretty}{Whether to pretty-print the JSON.}
}
\description{
Export failed checks to JSON
}
