% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{failed_checks}
\alias{failed_checks}
\title{Names of the failed checks}
\usage{
failed_checks(gp)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}
}
\value{
Names of the failed checks.
}
\description{
Names of the failed checks
}
\seealso{
Other API: \code{\link{checks}}, \code{\link{results}}
}

