% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{goodpractice-package}
\title{goodpractice: Advice on R Package Building}
\description{
Give advice about good practices when building R packages. Advice includes
functions and syntax to avoid, package structure, code complexity, code
formatting, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mangothecat/goodpractice}
  \item Report bugs at \url{https://github.com/mangothecat/goodpractice/issues}
}

}
\author{
\strong{Maintainer}: Daniel de Bortoli \email{daniel.debortoli@mango-solutions.com} (ddbortoli)

Authors:
\itemize{
  \item Gabor Csardi \email{csardi.gabor@gmail.com}
  \item Hannah Frick \email{hfrick@mango-solutions.com}
  \item Owen Jones \email{owenjonesuob@gmail.com} (owenjonesuob)
  \item Hannah Alexander \email{halexander@mango-solutions.com}
}

Other contributors:
\itemize{
  \item Mango Solutions (MangoTheCat) [copyright holder]
  \item Fabian Scheipl (fabian-s) [contributor]
}

}
