% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{goodpractice-package}
\alias{goodpractice-package}
\title{goodpractice: Advice on R Package Building}
\description{
Give advice about good practices when building R packages. Advice includes
functions and syntax to avoid, package structure, code complexity, code
formatting, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/goodpractice/}
  \item \url{https://github.com/ropensci-review-tools/goodpractice}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/goodpractice/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark@ropensci.org} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Authors:
\itemize{
  \item Karina Marks \email{karina.marks@ascent.io} (KarinaMarks)
  \item Daniel de Bortoli (ddbortoli)
  \item Gabor Csardi \email{csardi.gabor@gmail.com}
  \item Hannah Frick \email{hannah.frick@gmail.com}
  \item Owen Jones \email{owenjonesuob@gmail.com} (owenjonesuob)
  \item Hannah Alexander \email{halexander@mango-solutions.com}
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID}) (drmowinckels)
}

Other contributors:
\itemize{
  \item Ascent Digital Services UK Limited (MangoTheCat) [copyright holder]
  \item Ana Simmons \email{ana.simmons@ascent.io} (anasimmons) [contributor]
  \item Fabian Scheipl (fabian-s) [contributor]
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID}) (drmowinckels) [contributor]
}

}
