% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{goodpractice-package}
\title{goodpractice: Advice on R Package Building}
\description{
Give advice about good practices when building R packages. Advice includes
functions and syntax to avoid, package structure, code complexity, code
formatting, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mangothecat/goodpractice}
  \item Report bugs at \url{https://github.com/mangothecat/goodpractice/issues}
}

}
\author{
\strong{Maintainer}: Hannah Frick \email{hfrick@mango-solutions.com}

Authors:
\itemize{
  \item Gabor Csardi \email{csardi.gabor@gmail.com}
}

}
