% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{learn_skill_gp}
\alias{learn_skill_gp}
\title{Function for AI agents to run and learn how to use the 'goodpractice'
package.}
\usage{
learn_skill_gp()
}
\value{
The content of the file \code{system.file("skills",
"goodpractice4agents.md", package = "goodpractice"))}
}
\description{
You tell agents to directly use this function in order for them to learn the
skill. If you want to use the skill yourself to guide an agent, then use the
\link{use_skill_gp} function.
}
\examples{
\dontrun{
learn_gp_skill()
}
}
