% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_expressions.R
\name{package_collate}
\alias{package_collate}
\title{`Collate` field from `DESCRIPTION`}
\usage{
package_collate(path = ".")
}
\arguments{
\item{path}{Path to the package root.}
}
\value{
Character scalar or `NULL`.
}
\description{
`NULL` is returned if there is no such field.
}
\keyword{internal}
