% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.goodPractice}
\alias{print.goodPractice}
\title{Print goodpractice results}
\usage{
\method{print}{goodPractice}(x, groups = NULL, positions_limit = 5, ...)
}
\arguments{
\item{x}{Object of class \code{goodPractice}, as returned by \code{\link[=gp]{gp()}}.}

\item{groups}{Name of check groups for which to print results, as vector of
one or more of \code{\link[=all_check_groups]{all_check_groups()}}.}

\item{positions_limit}{How many positions to print at most.}

\item{...}{Unused, for compatibility with \code{\link[base:print]{base::print()}} generic method.}
}
\description{
Print goodpractice results
}
