% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.goodPractice}
\alias{print.goodPractice}
\title{Print goodpractice results}
\usage{
\method{print}{goodPractice}(x, positions_limit = 5, ...)
}
\arguments{
\item{x}{Object of class \code{goodPractice}, as returned by \code{\link[=gp]{gp()}}.}

\item{positions_limit}{how many positions to print at most.}
}
\description{
Print goodpractice results
}
