% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{results}
\alias{results}
\title{Return all check results in a data frame}
\usage{
results(gp)
}
\arguments{
\item{gp}{\code{\link{gp}} output.}
}
\value{
Data frame, with columns:
\item{check}{The name of the check.}
\item{result}{Logical, whether it has failed or not.}
}
\description{
Return all check results in a data frame
}
\examples{
path <- system.file("bad1", package = "goodpractice")
# run a subset of all checks available
g <- gp(path, checks = all_checks()[3:16])
results(g)
}
\seealso{
Other API: 
\code{\link{checks}()},
\code{\link{failed_checks}()}
}
\concept{API}
