% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{write_gp4agents}
\alias{write_gp4agents}
\title{Download a local copy of 'goodpractice4agents.md' to tell AI agents how to
use this package.}
\usage{
write_gp4agents(path = ".")
}
\arguments{
\item{path}{Local path where file should be extracted.}
}
\value{
(Invisibly) Full path to file.
\code{FALSE}.
}
\description{
The 'goodpractice4agents.md' file contains sufficient instructions for most
AI \%agents to edit package code so that it passes most 'goodpractice'
checks. The file can be directly edited and tweaked for personal use cases.
An agent can be instructed to simply "run the file goodpractice4agents.md".
The file can also be moved to any local agent's \code{skills/} directory to
call with a "skills" command. Alternatively, the file can be directly
downloaded from GitHub at
\url{https://github.com/ropensci-review-tools/goodpractice/tree/main/inst/agents/goodpractice4agents.md}.
}
\examples{
path <- file.path(tempdir(), "agents.md")
f <- write_gp4agents(path)
file.exists(f)
}
